requirements.txt
# Core Framework
streamlit==1.29.0
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0

# LangChain & LangGraph
langchain==0.1.0
langgraph==0.0.20
langchain-google-genai==0.0.6
langchain-community==0.0.10

# Vector Store & Embeddings
chromadb==0.4.22
sentence-transformers==2.3.1

# Data Sources (All FREE)
duckduckgo-search==4.1.1
wikipedia==1.4.0
yfinance==0.2.35
feedparser==6.0.10
beautifulsoup4==4.12.3
requests==2.31.0
lxml==5.1.0

# HTTP Client
httpx==0.26.0

# Document Generation
reportlab==4.0.9
python-docx==1.1.0
pypdf==3.17.4

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
loguru==0.7.2

# Development & Testing
pytest==7.4.3
black==23.12.1